file-scope: Move discard batch pipeline
Multi-file discard-to-batch support lived in discard.py even though it is
used by file-scope action dispatch rather than the discard command entry
point.

That kept batch-wide text hunk collection, aggregate reverse apply, and
multi-file result records mixed with single-command discard flow.

This commit moves the multi-file discard-to-batch pipeline into the
file-scope package and has multi-file action dispatch import it from there.

A following tests commit will pin the new boundary.
